+++ src/mockit/increment.ts
@@ -33,19 +33,7 @@ export default class ToIncrement extends Mockit<number | number[]> {
   }
   // generate
   public generate(options: TSuchInject): number | number[] {
-    let { $config, $length } = this.params;
-    if ($config && options.param?.$config) {
-      $config = {
-        ...$config,
-        ...options.param.$config,
-      };
-    }
-    if ($length && options.param?.$length) {
-      $length = {
-        ...$length,
-        ...options.param.$length,
-      };
-    }
+    const { $config, $length } = this.getCurrentParams(options);
     const { start, step } = ($config || {}) as {
       start?: number;
       step?: number;